requirements.txt
# Core
pytz==2021.3
Django==4.0.2
django-configurations==2.3.1
gunicorn==20.1.0
newrelic==7.2.4.171
django-cors-headers==3.11.0
celery==5.2.6
redis==4.2.2
flower==1.0.0

# Utils
django-environ==0.8.1
uuid==1.30

# For the persistence stores
psycopg2-binary==2.9.3
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.2.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.13.1
drf-writable-nested==0.6.3
Markdown==3.3.6
django-filter==21.1

# Developer Tools
ipdb==0.13.9
ipython==8.0.0
mkdocs==1.2.3
flake8==4.0.1
django-extensions==3.1.5
django_debug_toolbar==3.2.4

# Testing
mock==4.0.3
factory-boy==3.2.1
pytest==7.0.1
pytest-sugar==0.9.4
pytest-django==4.5.2
pytest-factoryboy==2.1.0
coverage==6.2
pytest-cov==3.0.0

# Static and Media Storage
django-storages==1.12.3
boto3==1.20.37

# Auth
mozilla-django-oidc==2.0.0
PyJWT==2.3.0

# Swagger
drf-spectacular==0.21.2
drf-spectacular-sidecar==2022.2.7

# Asynchronous server
channels==3.0.4
channels-redis==3.4.0
